backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
openai==1.58.1
